Fix bugs and continue work

+++ c_attempt/utils.h
@@ -16,12 +16,14 @@ static void write_int32_le(uint8_t *buf, int32_t value) {
     buf[3] = (value >> 24) & 0xFF;  // Byte 3 (MSB)
 }
 
-// Read 16-bit integer from buffer (little-endian)
 static int16_t read_int16_le(const uint8_t *buf) {
-    return buf[0] | (buf[1] << 8);
+    return (int16_t)((uint16_t)buf[0] | 
+                     ((uint16_t)buf[1] << 8));
 }
 
-// Read 32-bit integer from buffer (little-endian)
 static int32_t read_int32_le(const uint8_t *buf) {
-    return buf[0] | (buf[1] << 8) | (buf[2] << 16) | (buf[3] << 24);
+    return (int32_t)((uint32_t)buf[0] | 
+                     ((uint32_t)buf[1] << 8) | 
+                     ((uint32_t)buf[2] << 16) | 
+                     ((uint32_t)buf[3] << 24));
 }
